feat(KTreeList, KTruncate, KViewSwitcher): add types [KHCP-6974]
