+++ src/types/index.ts
@@ -31,3 +31,5 @@ export * from './toaster'
 export * from './method-badge'
 export * from './inline-edit'
 export * from './text-area'
+export * from './view-switcher'
+export * from './tree-list'
