+++ src/types/view-switcher.ts
@@ -1,0 +1,5 @@
+import { AnyElementOf } from '@/types/utils'
+
+export const SwitcherTypeArray = ['table', 'grid'] as const
+
+export type SwitcherType = AnyElementOf<typeof SwitcherTypeArray>
